python/requirements.txt
pyaudio==0.2.13
websockets==12.0
twilio==8.10.0
python-dotenv==1.0.0
aiohttp==3.9.1
googletrans==3.1.0a0
langdetect==1.0.9
indic-nlp-library==0.91
indic-transliteration==0.0.1 

